reestablish basic sphinx conf
