+++ doc/source/conf.py
@@ -16,10 +16,9 @@ import glob
 import shutil
 import sys
 import os
-import urllib
 
 sys.path.insert(0, os.path.abspath("."))
-from custom_directives import CustomGalleryItemDirective, fix_xgb_lgbm_docs
+from custom_directives import *
 from datetime import datetime
 
 # These lines added to enable Sphinx to work without installing Ray.
@@ -32,87 +31,9 @@ class ChildClassMock(mock.Mock):
         return mock.Mock
 
 
-MOCK_MODULES = [
-    "ax",
-    "ax.service.ax_client",
-    "blist",
-    "ConfigSpace",
-    "dask.distributed",
-    "gym",
-    "gym.spaces",
-    "horovod",
-    "horovod.runner",
-    "horovod.runner.common",
-    "horovod.runner.common.util",
-    "horovod.ray",
-    "horovod.ray.runner",
-    "horovod.ray.utils",
-    "hyperopt",
-    "hyperopt.hp"
-    "kubernetes",
-    "mlflow",
-    "modin",
-    "mxnet",
-    "mxnet.model",
-    "optuna",
-    "optuna.distributions",
-    "optuna.samplers",
-    "optuna.trial",
-    "psutil",
-    "ray._raylet",
-    "ray.core.generated",
-    "ray.core.generated.common_pb2",
-    "ray.core.generated.runtime_env_common_pb2",
-    "ray.core.generated.gcs_pb2",
-    "ray.core.generated.logging_pb2",
-    "ray.core.generated.ray.protocol.Task",
-    "ray.serve.generated",
-    "ray.serve.generated.serve_pb2",
-    "scipy.signal",
-    "scipy.stats",
-    "setproctitle",
-    "tensorflow_probability",
-    "tensorflow",
-    "tensorflow.contrib",
-    "tensorflow.contrib.all_reduce",
-    "tree",
-    "tensorflow.contrib.all_reduce.python",
-    "tensorflow.contrib.layers",
-    "tensorflow.contrib.rnn",
-    "tensorflow.contrib.slim",
-    "tensorflow.core",
-    "tensorflow.core.util",
-    "tensorflow.keras",
-    "tensorflow.python",
-    "tensorflow.python.client",
-    "tensorflow.python.util",
-    "torch",
-    "torch.distributed",
-    "torch.nn",
-    "torch.nn.parallel",
-    "torch.utils.data",
-    "torch.utils.data.distributed",
-    "wandb",
-    "zoopt",
-]
-
-CHILD_MOCK_MODULES = [
-    "pytorch_lightning",
-    "pytorch_lightning.accelerators",
-    "pytorch_lightning.plugins",
-    "pytorch_lightning.plugins.environments",
-    "pytorch_lightning.utilities",
-    "tensorflow.keras.callbacks",
-]
-
-import scipy.stats
-import scipy.linalg
-
 for mod_name in MOCK_MODULES:
     sys.modules[mod_name] = mock.Mock()
 
-# ray.rllib.models.action_dist.py and
-# ray.rllib.models.lstm.py will use tf.VERSION
 sys.modules["tensorflow"].VERSION = "9.9.9"
 
 for mod_name in CHILD_MOCK_MODULES:
@@ -151,13 +72,27 @@ extensions = [
     "versionwarning.extension",
     "sphinx_sitemap",
     "myst_parser",
+    # "myst_nb",
+    "sphinx.ext.doctest",
+    "sphinx.ext.coverage",
+    # "sphinx_external_toc",
 ]
+
+external_toc_exclude_missing = False
+# TODO activate later
+# external_toc_path = '_toc.yml'
+
+# There's a flaky autodoc import for "TensorFlowVariables" that fails depending on the doc structure / order
+# of imports.
+# autodoc_mock_imports = ["ray.experimental.tf_utils"]
+
+# This is used to suppress warnings about explicit "toctree" directives.
+# suppress_warnings = ["etoc.toctree"]
 
 versionwarning_admonition_type = "note"
 versionwarning_banner_title = "Join the Ray Discuss Forums!"
 
 FORUM_LINK = "https://discuss.ray.io"
-
 versionwarning_messages = {
     # Re-enable this after Ray Summit.
     # "latest": (
@@ -172,15 +107,19 @@ versionwarning_messages = {
 }
 
 versionwarning_body_selector = "#main-content"
+
 sphinx_gallery_conf = {
+    # Example sources are taken from these folders:
     "examples_dirs": [
-        "../examples",
+        "ray-core/_examples",
         "tune/_tutorials",
         "data/_examples",
-    ],  # path to example scripts
-    # path where to save generated examples
-    "gallery_dirs": ["auto_examples", "tune/tutorials", "data/examples"],
-    "ignore_pattern": "../examples/doc_code/",
+    ],
+    # and then generated into these respective target folders:
+    "gallery_dirs": [
+        "ray-core/examples", "tune/tutorials", "data/examples"
+    ],
+    "ignore_pattern": "ray-core/examples/doc_code/",
     "plot_gallery": "False",
     "min_reported_time": sys.maxsize,
     # "filename_pattern": "tutorial.py",
@@ -235,23 +174,22 @@ release = version
 # Usually you set "language" from the command line for these cases.
 language = None
 
-# There are two options for replacing |today|: either, you set today to some
-# non-false value, then it is used:
-# today = ''
-# Else, today_fmt is used as the format for a strftime call.
-# today_fmt = '%B %d, %Y'
-
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 exclude_patterns = ["_build"]
 exclude_patterns += sphinx_gallery_conf["examples_dirs"]
 
-# The reST default role (used for this markup: `text`) to use for all
-# documents.
-# default_role = None
+# If "DOC_LIB" is found, only build that top-level navigation item.
+build_one_lib = os.getenv("DOC_LIB")
 
-# If true, '()' will be appended to :func: etc. cross-reference text.
-# add_function_parentheses = True
+# TODO fix this later
+# All doc libs start with a "ray-" prefix.
+# all_toc_libs = [
+#     f.path for f in os.scandir(".") if f.is_dir() and "ray-" in f.path
+# ]
+# if build_one_lib and build_one_lib in all_toc_libs:
+#     all_toc_libs.remove(build_one_lib)
+#     exclude_patterns += all_toc_libs
 
 # If true, the current module name will be prepended to all description
 # unit titles (such as .. function::).
@@ -262,7 +200,7 @@ exclude_patterns += sphinx_gallery_conf["examples_dirs"]
 # show_authors = False
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = "pastie"
+pygments_style = "lovelace"
 
 # A list of ignored prefixes for module index sorting.
 # modindex_common_prefix = []
@@ -292,7 +230,7 @@ html_theme_options = {
     "use_issues_button": True,
     "use_edit_page_button": True,
     "path_to_docs": "doc/source",
-    "home_page_in_toc": True,
+    "home_page_in_toc": False,
     "show_navbar_depth": 0,
 }
 
@@ -301,7 +239,7 @@ html_theme_options = {
 
 # The name for this set of Sphinx documents.  If None, it defaults to
 # "<project> v<release> documentation".
-html_title = f"Ray v{release}"
+html_title = f"Ray {release}"
 
 # A shorter title for the navigation bar.  Default is the same as html_title.
 # html_short_title = None
@@ -355,9 +293,6 @@ html_static_path = ["_static"]
 # If true, "Created using Sphinx" is shown in the HTML footer. Default is True.
 # html_show_sphinx = True
 
-# If true, "(C) Copyright ..." is shown in the HTML footer. Default is True.
-# html_show_copyright = True
-
 # If true, an OpenSearch description file will be output, and all pages will
 # contain a <link> tag referring to it.  The value of this option must be the
 # base URL from which the finished HTML is served.
@@ -400,7 +335,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, "Ray.tex", "Ray Documentation", "The Ray Team", "manual"),
+    (master_doc, "Ray.tex", "Ray Documentation", author, "manual"),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -433,7 +368,6 @@ man_pages = [(master_doc, "ray", "Ray Documentation", [author], 1)]
 # man_show_urls = False
 
 # -- Options for Texinfo output -------------------------------------------
-
 # Grouping the document tree into Texinfo files. List of tuples
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
@@ -444,60 +378,20 @@ texinfo_documents = [
         "Ray Documentation",
         author,
         "Ray",
-        "One line description of project.",
+        "Ray provides a simple, universal API for building distributed applications.",
         "Miscellaneous",
     ),
 ]
 
-# Documents to append as an appendix to all manuals.
-# texinfo_appendices = []
-
-# If false, no module index is generated.
-# texinfo_domain_indices = True
-
-# How to display URL addresses: 'footnote', 'no', or 'inline'.
-# texinfo_show_urls = 'footnote'
-
-# If true, do not generate a @detailmenu in the "Top" node's menu.
-# texinfo_no_detailmenu = False
-
-# pcmoritz: To make the following work, you have to run
-# sudo pip install recommonmark
-
 # Python methods should be presented in source code order
 autodoc_member_order = "bysource"
-
-# Taken from https://github.com/edx/edx-documentation
-FEEDBACK_FORM_FMT = "https://github.com/ray-project/ray/issues/new?title={title}&labels=docs&body={body}"
-
-
-def feedback_form_url(project, page):
-    """Create a URL for feedback on a particular page in a project."""
-    return FEEDBACK_FORM_FMT.format(
-        title=urllib.parse.quote(
-            "[docs] Issue on `{page}.rst`".format(page=page)),
-        body=urllib.parse.quote(
-            "# Documentation Problem/Question/Comment\n"
-            "<!-- Describe your issue/question/comment below. -->\n"
-            "<!-- If there are typos or errors in the docs, feel free to create a pull-request. -->\n"
-            "\n\n\n\n"
-            "(Created directly from the docs)\n"),
-    )
-
-
-def update_context(app, pagename, templatename, context, doctree):
-    """Update the page rendering context to include ``feedback_form_url``."""
-    context["feedback_form_url"] = feedback_form_url(app.config.project,
-                                                     pagename)
-
-
-# see also http://searchvoidstar.tumblr.com/post/125486358368/making-pdfs-from-markdown-on-readthedocsorg-using
 
 
 def setup(app):
     app.connect("html-page-context", update_context)
+    # Custom CSS
     app.add_css_file("css/custom.css")
-    # Custom directives
+    # Custom Sphinx directives
     app.add_directive("customgalleryitem", CustomGalleryItemDirective)
-    # Custom connects
+    # Custom docstring processor
     app.connect("autodoc-process-docstring", fix_xgb_lgbm_docs)
